--- doomarena/osworld/src/doomarena/osworld/attacks/attack_config/intent_click_tgt_OK.json
@@ -1,7 +1,0 @@
-{
-    "bottom": "True",
-    "notice": "OK",
-    "attack_string": "instruct click tgt",
-    "prefix": "intent",
-    "adv_text": "True"
-}
